admin_panel/requirements.txt
asgiref==3.5.1
beautifulsoup4==4.11.1
Django==4.0.4
django-bootstrap4==22.1
django-fontawesome-5==1.0.18
django-livesync==0.5
djangorestframework==3.13.1
jquery.javascript==1.0.0
multipart==0.2.4
python-gettext==4.0
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
transaction==3.0.1
watchdog==2.1.7
websocket-client==1.3.2
zope.browser==2.4
zope.browserpage==4.4.0
zope.component==5.0.1
zope.configuration==4.4.1
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.exceptions==4.5
zope.hookable==5.1.0
zope.i18n==4.9.0
zope.i18nmessageid==5.0.1
zope.interface==5.4.0
zope.location==4.2
zope.pagetemplate==4.6.0
zope.proxy==4.5.0
zope.publisher==6.1.0
zope.schema==6.2.0
zope.security==5.3
zope.tal==4.5
zope.tales==5.1
zope.traversing==4.4.1
zope.viewlet==4.3
